move state management related to planets from App to PlanetProvider component
